+++ use-cases/conference-scheduling/src/main/java/org/acme/conferencescheduling/solver/ConferenceSchedulingConstraintProvider.java
@@ -26,42 +26,8 @@ import ai.timefold.solver.service.definition.api.description.ConstraintInfo;
 import org.acme.conferencescheduling.domain.ConferenceConstraintProperties;
 import org.acme.conferencescheduling.domain.Speaker;
 import org.acme.conferencescheduling.domain.Talk;
-import org.acme.conferencescheduling.domain.justification.CrowdControlTalkNotPairedJustification;
-import org.acme.conferencescheduling.domain.justification.MissingPreferredRoomTagsForSpeakersJustification;
-import org.acme.conferencescheduling.domain.justification.MissingPreferredRoomTagsForTalkJustification;
-import org.acme.conferencescheduling.domain.justification.MissingPreferredTimeslotTagsForSpeakersJustification;
-import org.acme.conferencescheduling.domain.justification.MissingPreferredTimeslotTagsForTalkJustification;
-import org.acme.conferencescheduling.domain.justification.MissingRequiredRoomTagsForSpeakersJustification;
-import org.acme.conferencescheduling.domain.justification.MissingRequiredRoomTagsForTalkJustification;
-import org.acme.conferencescheduling.domain.justification.MissingRequiredTimeslotTagsForSpeakersJustification;
-import org.acme.conferencescheduling.domain.justification.MissingRequiredTimeslotTagsForTalkJustification;
-import org.acme.conferencescheduling.domain.justification.MutuallyExclusiveTalksOverlappingJustification;
-import org.acme.conferencescheduling.domain.justification.PopularTalkInSmallerRoomJustification;
-import org.acme.conferencescheduling.domain.justification.ProhibitedRoomTagsForSpeakersJustification;
-import org.acme.conferencescheduling.domain.justification.ProhibitedRoomTagsForTalkJustification;
-import org.acme.conferencescheduling.domain.justification.ProhibitedTimeslotTagsForSpeakersJustification;
-import org.acme.conferencescheduling.domain.justification.ProhibitedTimeslotTagsForTalkJustification;
-import org.acme.conferencescheduling.domain.justification.RelatedTalksNotOnSameDayJustification;
-import org.acme.conferencescheduling.domain.justification.RoomUnavailableAtTalkTimeslotJustification;
-import org.acme.conferencescheduling.domain.justification.SharedContentAudienceLevelFlowViolationJustification;
-import org.acme.conferencescheduling.domain.justification.SpeakerAssignedToOverlappingTalksJustification;
-import org.acme.conferencescheduling.domain.justification.SpeakerConsecutiveTalksPauseTooShortJustification;
-import org.acme.conferencescheduling.domain.justification.SpeakerMakespanTooLongJustification;
-import org.acme.conferencescheduling.domain.justification.SpeakerUnavailableAtTalkTimeslotJustification;
-import org.acme.conferencescheduling.domain.justification.TalkScheduledBeforePrerequisiteTalkJustification;
-import org.acme.conferencescheduling.domain.justification.TalksOverlappingInSameRoomJustification;
-import org.acme.conferencescheduling.domain.justification.TalksWithDifferentAudienceLevelInSameTimeslotJustification;
-import org.acme.conferencescheduling.domain.justification.TalksWithSameAudienceTypeInSameTimeslotJustification;
-import org.acme.conferencescheduling.domain.justification.TalksWithSameContentOverlappingJustification;
-import org.acme.conferencescheduling.domain.justification.TalksWithSameLanguageInSameTimeslotJustification;
-import org.acme.conferencescheduling.domain.justification.TalksWithSameSectorOverlappingJustification;
-import org.acme.conferencescheduling.domain.justification.TalksWithSameThemeTrackAndAudienceTypeOverlappingJustification;
-import org.acme.conferencescheduling.domain.justification.TalksWithSameThemeTrackInDifferentRoomsJustification;
-import org.acme.conferencescheduling.domain.justification.TalksWithSameThemeTrackOverlappingJustification;
-import org.acme.conferencescheduling.domain.justification.UndesiredRoomTagsForSpeakersJustification;
-import org.acme.conferencescheduling.domain.justification.UndesiredRoomTagsForTalkJustification;
-import org.acme.conferencescheduling.domain.justification.UndesiredTimeslotTagsForSpeakersJustification;
-import org.acme.conferencescheduling.domain.justification.UndesiredTimeslotTagsForTalkJustification;
+import org.acme.conferencescheduling.domain.justification.ConferenceSchedulingJustification;
+import org.acme.conferencescheduling.domain.justification.ConferenceSchedulingJustification.*;
 
 /**
  * Provides the constraints for the conference scheduling problem.
@@ -160,7 +126,7 @@ public class ConferenceSchedulingConstraintProvider implements ConstraintProvide
                         containedIn(Talk::getTimeslot, Speaker::unavailableTimeslots))
                 .penalize(HardMediumSoftScore.ofHard(100), (talk, speaker) -> talk.getDurationInMinutes())
                 .justifyWith(
-                        (talk, speaker, score) -> SpeakerUnavailableAtTalkTimeslotJustification.of(talk, speaker))
+                        (talk, speaker, score) -> ConferenceSchedulingJustification.SpeakerUnavailableAtTalkTimeslotJustification.of(talk, speaker))
                 .asConstraint(new ConstraintInfo(ConferenceConstraintProperties.SPEAKER_UNAVAILABLE_TIMESLOT,
                         ConferenceConstraintProperties.SPEAKER_UNAVAILABLE_TIMESLOT,
                         "A talk must not be scheduled in a timeslot when one of its speakers is unavailable.",
