+++ use-cases/conference-scheduling/src/test/java/org/acme/conferencescheduling/solver/SolverManagerTest.java
@@ -14,7 +14,7 @@ import org.acme.conferencescheduling.domain.ConferenceSchedule;
 import org.acme.conferencescheduling.dto.ConferenceScheduleInputMetrics;
 import org.acme.conferencescheduling.dto.ConferenceScheduleOutputMetrics;
 import org.acme.conferencescheduling.service.ConferenceScheduleModelConvertor;
-import org.acme.conferencescheduling.support.SolverTestDataFactory;
+import org.acme.conferencescheduling.support.TestHelper;
 import org.junit.jupiter.api.Test;
 
 import io.quarkus.test.junit.QuarkusTest;
@@ -30,7 +30,7 @@ class SolverManagerTest {
 
     @Test
     void solve() throws ExecutionException, InterruptedException {
-        var input = SolverTestDataFactory.createProblem();
+        var input = TestHelper.createProblem();
 
         ConferenceSchedule problem = modelConvertor.toSolverModel(input, ModelConfig.empty(), Optional.empty());
